requirements.txt
torch>=2.1
transformers>=4.40
optuna>=3.5
rasterio>=1.3
laspy>=2.5
pyproj>=3.6
scikit-learn>=1.3
scikit-image>=0.22
scipy>=1.11
opencv-python>=4.9
pandas>=2.1
matplotlib>=3.8
seaborn>=0.13
Pillow>=10
tqdm>=4.66
numpy>=1.26
PyYAML>=6.0
